requirements.txt
# Core Flask dependencies
Flask==3.0.3
Werkzeug>=2.0.0

# Data handling
PyYAML>=6.0
requests>=2.28.0

# AI Integration
google-generativeai>=0.3.0

# Image processing and QR codes
Pillow==10.4.0
qrcode[pil]==7.4.2

# Bluetooth Label Printing (Niimbot printers)
bleak>=0.21.0  # Bluetooth Low Energy support for Niimbot printers

# Optional: For enhanced database features
# sqlite3 is included with Python by default

# Optional: Advanced image processing (requires build tools)
# pycairo>=1.20.0  # Commented out - requires C compiler, not needed for basic Niimbot functionality

# Development dependencies (uncomment for development)
# flask-debugtoolbar
# pytest
# pytest-flask